# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/v1/properties/": {
            "post": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Create a property",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Properties"
                ],
                "summary": "Create a property",
                "parameters": [
                    {
                        "description": "Create a property",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Aliizi83_sample-golang-project_src_api_dto.CreatePropertyRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Property response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_Aliizi83_sample-golang-project_src_api_helpers.BaseHttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/github_com_Aliizi83_sample-golang-project_src_api_dto.PropertyResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_Aliizi83_sample-golang-project_src_api_helpers.BaseHttpResponse"
                        }
                    }
                }
            }
        },
        "/v1/properties/get-by-filters": {
            "post": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Get Properties",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Properties"
                ],
                "summary": "Get Properties",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Aliizi83_sample-golang-project_src_domain_filters.PaginationInputWithFilter"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Property response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_Aliizi83_sample-golang-project_src_api_helpers.BaseHttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/github_com_Aliizi83_sample-golang-project_src_domain_filters.PagedList-github_com_Aliizi83_sample-golang-project_src_api_dto_PropertyResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_Aliizi83_sample-golang-project_src_api_helpers.BaseHttpResponse"
                        }
                    }
                }
            }
        },
        "/v1/properties/{id}": {
            "get": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Get a property",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Properties"
                ],
                "summary": "Get a property",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "property response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_Aliizi83_sample-golang-project_src_api_helpers.BaseHttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/github_com_Aliizi83_sample-golang-project_src_api_dto.PropertyResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_Aliizi83_sample-golang-project_src_api_helpers.BaseHttpResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Update a property",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Properties"
                ],
                "summary": "Update a property",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update a property",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Aliizi83_sample-golang-project_src_api_dto.UpdatePropertyRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Property response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_Aliizi83_sample-golang-project_src_api_helpers.BaseHttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/github_com_Aliizi83_sample-golang-project_src_api_dto.PropertyResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_Aliizi83_sample-golang-project_src_api_helpers.BaseHttpResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Delete a property",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Properties"
                ],
                "summary": "Delete a property",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "response",
                        "schema": {
                            "$ref": "#/definitions/github_com_Aliizi83_sample-golang-project_src_api_helpers.BaseHttpResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_Aliizi83_sample-golang-project_src_api_helpers.BaseHttpResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_Aliizi83_sample-golang-project_src_api_dto.CreatePropertyRequest": {
            "type": "object",
            "required": [
                "categoryId",
                "name"
            ],
            "properties": {
                "categoryId": {
                    "type": "integer"
                },
                "dataType": {
                    "type": "string",
                    "maxLength": 15
                },
                "description": {
                    "type": "string",
                    "maxLength": 1000
                },
                "icon": {
                    "type": "string",
                    "maxLength": 1000
                },
                "name": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 3
                },
                "unit": {
                    "type": "string",
                    "maxLength": 15
                }
            }
        },
...
        "github_com_Aliizi83_sample-golang-project_src_api_dto.PropertyCategoryResponse": {
            "type": "object",
            "properties": {
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "properties": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_Aliizi83_sample-golang-project_src_api_dto.PropertyResponse"
                    }
                }
            }
        },
        "github_com_Aliizi83_sample-golang-project_src_api_dto.PropertyResponse": {
            "type": "object",
            "properties": {
                "category": {
                    "$ref": "#/definitions/github_com_Aliizi83_sample-golang-project_src_api_dto.PropertyCategoryResponse"
                },
                "dataType": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "unit": {
                    "type": "string"
                }
            }
        },
...
        "github_com_Aliizi83_sample-golang-project_src_api_dto.UpdatePropertyRequest": {
            "type": "object",
            "properties": {
                "categoryId": {
                    "type": "integer"
                },
                "dataType": {
                    "type": "string",
                    "maxLength": 15
                },
                "description": {
                    "type": "string",
                    "maxLength": 1000
                },
                "icon": {
                    "type": "string",
                    "maxLength": 1000
                },
                "name": {
                    "type": "string"
                },
                "unit": {
                    "type": "string",
                    "maxLength": 15
                }
            }
        },
        "github_com_Aliizi83_sample-golang-project_src_api_helpers.BaseHttpResponse": {
            "type": "object",
            "properties": {
                "error": {},
                "result": {},
                "resultCode": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                },
                "validationErrors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_Aliizi83_sample-golang-project_src_api_validations.ValidationError"
                    }
                }
            }
        },
        "github_com_Aliizi83_sample-golang-project_src_api_validations.ValidationError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "property": {
                    "type": "string"
                },
                "tag": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "github_com_Aliizi83_sample-golang-project_src_domain_filters.Filter": {
            "type": "object",
            "properties": {
                "filterType": {
                    "type": "string"
                },
                "from": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "github_com_Aliizi83_sample-golang-project_src_domain_filters.PagedList-github_com_Aliizi83_sample-golang-project_src_api_dto_PropertyResponse": {
            "type": "object",
            "properties": {
                "hasNextPage": {
                    "type": "boolean"
                },
                "hasPreviousPage": {
                    "type": "boolean"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_Aliizi83_sample-golang-project_src_api_dto.PropertyResponse"
                    }
                },
                "pageNumber": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                },
                "totalRows": {
                    "type": "integer"
                }
            }
        },
        "github_com_Aliizi83_sample-golang-project_src_domain_filters.PaginationInputWithFilter": {
            "type": "object",
            "properties": {
                "filters": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/github_com_Aliizi83_sample-golang-project_src_domain_filters.Filter"
                    }
                },
                "pageNumber": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "sorts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_Aliizi83_sample-golang-project_src_domain_filters.Sort"
                    }
                }
            }
        },
        "github_com_Aliizi83_sample-golang-project_src_domain_filters.Sort": {
            "type": "object",
            "properties": {
                "columnId": {
                    "type": "string"
                },
                "sort": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "AuthBearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}